requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
PyYAML
pyiqa
loguru
progress
tensorboard
tqdm
einops
timm
basicsr
openpyxl
python-docx
